reducing complexity in checkUserExistense
