+++ src/middlewares/checkUserExistence.js
@@ -1,5 +1,5 @@
-import models from '../models';
 import userExist from '../services/findUser';
+import register from '../helper/googleRegister'
 const checkUserExistance=async(req,res,next)=>{
         let user;
         if(req.user){ 
@@ -9,22 +9,8 @@ const checkUserExistance=async(req,res,next)=>{
         }
         let account= await userExist(user.email);
         if(!account){
-            const User={first_name:user.given_name,
-                last_name:user.family_name,
-                username:user.displayName,email:user.email,
-                language:user.language,
-                profile_picture:user.photos[0].value,varified:user.verified,
-            };
-            try{
-                account=await models.User.create(User);
-                next()
-            }catch(err){
-                return res.status(500).send({
-                    message:"failed signing up",
-                    error:err.message
-                })
-            }
-        
+           await register(user);
+           next();
         }else{
             next();
         }
